Summary: billing code D2394: resin-based composite - four or more surfaces, posterior
'is about' some 
    ('direct restoration procedure'
     and (has_specified_input some 'resin dental restoration material')
     and (has_specified_input some 'Posterior Tooth'))

'is about' some 
    ('resin dental restoration material'
     and (has_specified_input some 'Posterior Tooth')
     and ('is dental restoration of' min 4 'Surface enamel of tooth'))

Unsat because  'resin dental restoration material' (material entity) and  (has_specified_input some 'Posterior Tooth') (planned process)

Same for the other one.

I'm repairing it by changing the second axiom to drop the the   " and (has_specified_input some 'Posterior Tooth')"

I added a few more disjoints and some others came up. All of them are of the form

subclass of billing code ....
'is dental restoration of' ...

Which fails because billing code is gdc and domain of is dental restoration of is material entity.
I'm fixing them to have the second clause wrapped in an 'is about'

****************************************************************

Added 

Role: target of dental procedure role
Annnotation property: duplicated